+++ demo/nextjs/pages/thumbnail-render-thumbnail-item-dynamic-document/index.tsx
@@ -35,7 +35,9 @@ const IndexPage = () => {
         </div>
     );
 
-    const thumbnailPluginInstance = thumbnailPlugin();
+    const thumbnailPluginInstance = thumbnailPlugin({
+        thumbnailWidth: 150,
+    });
     const { Thumbnails } = thumbnailPluginInstance;
 
     return (