+++ packages/thumbnail/__tests__/thumbnailWidth.test.tsx
@@ -1,33 +1,14 @@
 import * as React from 'react';
-import { fireEvent, render, screen, waitForElementToBeRemoved } from '@testing-library/react';
-import { classNames, Viewer } from '@react-pdf-viewer/core';
+import { render, screen, waitForElementToBeRemoved } from '@testing-library/react';
+import { Viewer } from '@react-pdf-viewer/core';
 
 import { mockIsIntersecting } from '../../../test-utils/mockIntersectionObserver';
 import { mockResize } from '../../../test-utils/mockResizeObserver';
-import { thumbnailPlugin, RenderThumbnailItemProps } from '../src';
+import { thumbnailPlugin } from '../src';
 
-const TestRenderThumbnailItem: React.FC<{
+const TestThumbnailWidth: React.FC<{
     fileUrl: Uint8Array;
 }> = ({ fileUrl }) => {
-    const renderThumbnailItem = (props: RenderThumbnailItemProps) => (
-        <div
-            key={props.pageIndex}
-            className={classNames({
-                'custom-thumbnail-item': true,
-                'custom-thumbnail-item--selected': props.pageIndex === props.currentPage,
-            })}
-            data-testid={`thumbnail-${props.pageIndex}`}
-            style={{
-                backgroundColor: props.pageIndex === props.currentPage ? 'rgba(0, 0, 0, 0.3)' : '#fff',
-                cursor: 'pointer',
-                padding: '0.5rem',
-            }}
-            onClick={props.onJumpToPage}
-        >
-            {props.renderPageThumbnail}
-        </div>
-    );
-
     const thumbnailPluginInstance = thumbnailPlugin({
         thumbnailWidth: 150,
     });
@@ -40,16 +21,17 @@ const TestRenderThumbnailItem: React.FC<{
                 display: 'flex',
                 height: '50rem',
                 width: '50rem',
+                margin: '1rem auto',
             }}
         >
             <div
                 style={{
                     borderRight: '1px solid rgba(0, 0, 0, 0.3)',
                     overflow: 'auto',
-                    width: '30%',
+                    width: '15rem',
                 }}
             >
-                <Thumbnails renderThumbnailItem={renderThumbnailItem} />
+                <Thumbnails />
             </div>
             <div style={{ flex: 1 }}>
                 <Viewer fileUrl={fileUrl} plugins={[thumbnailPluginInstance]} />
@@ -58,15 +40,15 @@ const TestRenderThumbnailItem: React.FC<{
     );
 };
 
-test('Test renderThumbnailItem option', async () => {
+test('Test thumbnailWidth option', async () => {
     const { findByLabelText, findByTestId, getByTestId } = render(
-        <TestRenderThumbnailItem fileUrl={global['__OPEN_PARAMS_PDF__']} />
+        <TestThumbnailWidth fileUrl={global['__OPEN_PARAMS_PDF__']} />
     );
 
     const viewerEle = getByTestId('core__viewer');
     mockIsIntersecting(viewerEle, true);
-    viewerEle['__jsdomMockClientHeight'] = 559;
-    viewerEle['__jsdomMockClientWidth'] = 800;
+    viewerEle['__jsdomMockClientHeight'] = 558;
+    viewerEle['__jsdomMockClientWidth'] = 798;
 
     // Wait until the document is loaded completely
     await waitForElementToBeRemoved(() => screen.getByTestId('core__doc-loading'));
@@ -83,8 +65,8 @@ test('Test renderThumbnailItem option', async () => {
     pagesContainer.getBoundingClientRect = jest.fn(() => ({
         x: 0,
         y: 0,
-        height: 559,
-        width: 800,
+        height: 558,
+        width: 798,
         top: 0,
         right: 0,
         bottom: 0,
@@ -97,27 +79,27 @@ test('Test renderThumbnailItem option', async () => {
     mockIsIntersecting(thumbnailsListContainer, true);
 
     const thumbnailsContainer = await findByTestId('thumbnail__list');
-    expect(thumbnailsContainer.querySelectorAll('.custom-thumbnail-item').length).toEqual(8);
+    expect(thumbnailsContainer.querySelectorAll('.rpv-thumbnail__item').length).toEqual(8);
 
-    // Find the second thumbnail
-    let secondThumbnail = await findByTestId('thumbnail-1');
-    expect(secondThumbnail).toHaveClass('custom-thumbnail-item');
+    // Find the first thumbnail
+    let firstThumbnailContainer = await findByTestId('thumbnail__container-0');
+    mockIsIntersecting(firstThumbnailContainer, true);
 
-    // Scroll to the second page
-    fireEvent.scroll(pagesContainer, {
-        target: {
-            scrollTop: 658,
-        },
-    });
-
-    await findByTestId('core__text-layer-2');
+    const firstThumbnailImage = await findByLabelText('Thumbnail of page 1');
+    let src = firstThumbnailImage.getAttribute('src');
+    expect(src.substring(0, 100)).toEqual(
+        'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAJYAAADICAYAAAAKhRhlAAAABmJLR0QA/wD/AP+gvaeTAAASp0lEQV'
+    );
+    expect(src.length).toEqual(6490);
+    expect(firstThumbnailImage.getAttribute('width')).toEqual('150px');
+    expect(firstThumbnailImage.getAttribute('height')).toEqual('200px');
 
     // Wait until the second thumbnail is rendered
     let secondThumbnailContainer = await findByTestId('thumbnail__container-1');
     mockIsIntersecting(secondThumbnailContainer, true);
 
     const secondThumbnailImage = await findByLabelText('Thumbnail of page 2');
-    const src = secondThumbnailImage.getAttribute('src');
+    src = secondThumbnailImage.getAttribute('src');
     expect(src.substring(0, 100)).toEqual(
         'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAJYAAADICAYAAAAKhRhlAAAABmJLR0QA/wD/AP+gvaeTAAAgAElEQV'
     );